Reduced particleID data structure size

+++ src/Client/ParticleID.hpp
@@ -3,7 +3,7 @@
 #include <cstdint>
 
 // THE ORDER IN WHICH THE PARTICLES ARE DEFINED MUST MATCH THE ORDER IN WHICH THEY ARE ASSIGNED IN ParticleTexture.cpp:populateUVQuadArray().
-enum class ParticleID : unsigned int {
+enum class ParticleID : uint8_t {
 	DIRT,
 	STONE
 };
